Some CAC / PIV cards donot support Discovery Object
Some CAC card return '6A80` Incorrect parameters in APDU
when trying to read the Discovery object.
If it fails other then not found, then we can not use the Discovery
object to test for the active AID.

The test is done in piv_match_card jsut after doing a SELECT AID for the PIV.
and set CI_DISCOVERY_USELESS if needed.

piv_card_reader_lock_obtained will then not use teh Discovery object.

 On branch piv-aid-discovery
 Changes to be committed:
	modified:   card-piv.c